fix(backend): address critical review findings
1. OnConflict unique constraint now includes organization_id
   (prevents cross-org grant collisions)
2. SQL sub-queries use parameterized grant.TypePod/TypeRunner/TypeRepository
   constants instead of hardcoded 'pod'/'runner'/'repository' strings
3. resource_grants.go (339 lines) split into pod_grants.go (93),
   runner_grants.go (122), repository_grants.go (122) — all under 200-line limit
4. Grant cleanup on resource deletion: TerminatePod, DeleteRunner,
   DeleteRepository now call grantService.CleanupByResource()
5. Added idx_resource_grants_lookup index for efficient sub-query joins
