+++ backend/internal/infra/gitprovider_repo.go
@@ -5,6 +5,7 @@ import (
 	"time"
 
 	"github.com/anthropics/agentsmesh/backend/internal/domain/gitprovider"
+	"github.com/anthropics/agentsmesh/backend/internal/domain/grant"
 	"github.com/anthropics/agentsmesh/backend/internal/domain/ticket"
 	"gorm.io/gorm"
 )
@@ -93,8 +94,8 @@ func (r *gitProviderRepo) ListByOrganizationForUser(ctx context.Context, orgID i
 	var repos []*gitprovider.Repository
 	err := r.db.WithContext(ctx).
 		Where("organization_id = ? AND is_active = ? AND deleted_at IS NULL", orgID, true).
-		Where("(visibility = 'organization' OR (visibility = 'private' AND imported_by_user_id = ?) OR id::text IN (SELECT resource_id FROM resource_grants WHERE resource_type = 'repository' AND user_id = ? AND organization_id = ?))",
-			userID, userID, orgID).
+		Where("(visibility = 'organization' OR (visibility = 'private' AND imported_by_user_id = ?) OR id::text IN (SELECT resource_id FROM resource_grants WHERE resource_type = ? AND user_id = ? AND organization_id = ?))",
+			userID, grant.TypeRepository, userID, orgID).
 		Order("created_at DESC").
 		Find(&repos).Error
 	return repos, err
