+++ backend/internal/infra/runner_repo.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"time"
 
+	"github.com/anthropics/agentsmesh/backend/internal/domain/grant"
 	"github.com/anthropics/agentsmesh/backend/internal/domain/runner"
 	"gorm.io/gorm"
 )
@@ -88,14 +89,14 @@ func (r *runnerRepository) Delete(ctx context.Context, runnerID int64) error {
 
 // --- Runner Queries ---
 
-// visibilityWithGrantsFilter returns a SQL fragment that includes visibility checks
-// plus resource_grants access. Bind params: (userID, userID, orgID).
-const visibilityWithGrantsFilter = "(visibility = 'organization' OR (visibility = 'private' AND registered_by_user_id = ?) OR id::text IN (SELECT resource_id FROM resource_grants WHERE resource_type = 'runner' AND user_id = ? AND organization_id = ?))"
+// visibilityWithGrantsFilter is a SQL fragment for visibility + grant access.
+// Bind params: (userID, grantResourceType, userID, orgID).
+const visibilityWithGrantsFilter = "(visibility = 'organization' OR (visibility = 'private' AND registered_by_user_id = ?) OR id::text IN (SELECT resource_id FROM resource_grants WHERE resource_type = ? AND user_id = ? AND organization_id = ?))"
 
 func (r *runnerRepository) ListByOrg(ctx context.Context, orgID, userID int64) ([]*runner.Runner, error) {
 	var runners []*runner.Runner
 	if err := r.db.WithContext(ctx).
-		Where("organization_id = ? AND "+visibilityWithGrantsFilter, orgID, userID, userID, orgID).
+		Where("organization_id = ? AND "+visibilityWithGrantsFilter, orgID, userID, grant.TypeRunner, userID, orgID).
 		Find(&runners).Error; err != nil {
 		return nil, err
 	}
@@ -106,7 +107,7 @@ func (r *runnerRepository) ListAvailable(ctx context.Context, orgID, userID int6
 	var runners []*runner.Runner
 	if err := r.db.WithContext(ctx).
 		Where("organization_id = ? AND status = ? AND is_enabled = ? AND current_pods < max_concurrent_pods AND "+visibilityWithGrantsFilter,
-			orgID, runner.RunnerStatusOnline, true, userID, userID, orgID).
+			orgID, runner.RunnerStatusOnline, true, userID, grant.TypeRunner, userID, orgID).
 		Find(&runners).Error; err != nil {
 		return nil, err
 	}
@@ -117,7 +118,7 @@ func (r *runnerRepository) ListAvailableOrdered(ctx context.Context, orgID, user
 	var runners []*runner.Runner
 	if err := r.db.WithContext(ctx).
 		Where("organization_id = ? AND status = ? AND is_enabled = ? AND current_pods < max_concurrent_pods AND "+visibilityWithGrantsFilter,
-			orgID, runner.RunnerStatusOnline, true, userID, userID, orgID).
+			orgID, runner.RunnerStatusOnline, true, userID, grant.TypeRunner, userID, orgID).
 		Order("current_pods ASC").
 		Find(&runners).Error; err != nil {
 		return nil, err
@@ -129,7 +130,7 @@ func (r *runnerRepository) ListAvailableForAgent(ctx context.Context, orgID, use
 	var runners []*runner.Runner
 	if err := r.db.WithContext(ctx).
 		Where("organization_id = ? AND status = ? AND is_enabled = ? AND current_pods < max_concurrent_pods AND available_agents @> ? AND "+visibilityWithGrantsFilter,
-			orgID, runner.RunnerStatusOnline, true, agentJSON, userID, userID, orgID).
+			orgID, runner.RunnerStatusOnline, true, agentJSON, userID, grant.TypeRunner, userID, orgID).
 		Order("current_pods ASC").
 		Find(&runners).Error; err != nil {
 		return nil, err
